requirements.txt
requests
pandas
python-dotenv
vaderSentiment
streamlit
tweepy
logging